MH3-22330 - Add generic touch controller driver generator
This component replaces the generic controller component in input. This
driver generator will generate a driver that allows initialization of
the controller using I2C commands and parameters and allows reading and
parsing of touch data of a certain format. The driver generator allows
users to specify the index, mask and offset of the touch event and
position to extract the event and position data. The event and position
data are sent to the input system service for processing.

This initial version of the driver generator has been tested to generate
a basic driver that works with the FocalTect FT5306 touch controller on
the Mikroe 5" Capacitive display package.